all: rename services to cmd
I like this name better
