+++ README.md
@@ -26,11 +26,11 @@ Working:
  - Go gRPC server
     ```
     $ docker run --rm -d --name postgres -p 5432:5432 -e POSTGRES_HOST_AUTH_METHOD=trust postgres
-    $ bazel run //service/go-server:go-server.binary -- --postgres-url postgresql://postgres@localhost:5432/postgres?sslmode=disable
-    $ bazel run //service/go-server:publish
+    $ bazel run //cmd/go-server:go-server.binary -- --postgres-url postgresql://postgres@localhost:5432/postgres?sslmode=disable
+    $ bazel run //cmd/go-server:publish
     ```
  - Go gRPC client
     ```
-    $ bazel run //service/go-client:go-client.binary
-    $ bazel run //service/go-client:publish
+    $ bazel run //cmd/go-client:go-client.binary
+    $ bazel run //cmd/go-client:publish
     ```
