+++ cmd/go-server/users/users_test.go
@@ -19,8 +19,8 @@ import (
 	"google.golang.org/grpc/status"
 	logrusadapter "logur.dev/adapter/logrus"
 
+	"github.com/johanbrandhorst/bazel-mono/cmd/go-server/users"
 	pbUsers "github.com/johanbrandhorst/bazel-mono/proto/myorg/users/v1"
-	"github.com/johanbrandhorst/bazel-mono/service/go-server/users"
 )
 
 var (
